refactored data loading code support for JINX data format in addition to Parquet
