+++ maester/datasets/formats/__init__.py
@@ -1,0 +1,2 @@
+from .jinx import JinxDataset
+from .parquet import ParquetDataset
